requirements.txt
openai
mcp
numpy
aiosqlite
pandas
python-dotenv